requirements.txt
anyio==3.6.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
dbus-python==1.2.16
distro-info==1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==2.10
iotop==0.6
markdown-it-py==2.1.0
mdurl==0.1.2
nala==0.13.0
netifaces==0.10.9
OPi.GPIO==0.5.2
pexpect==4.8.0
pycurl==7.43.0.6
Pygments==2.14.0
PyGObject==3.38.0
python-apt==2.2.1
python-debian==0.1.49
PyYAML==5.3.1
requests==2.28.1
rfc3986==1.4.0
rich==13.2.0
six==1.16.0
sniffio==1.1.0
tomli==2.0.1
typer==0.7.0
typing-extensions==4.4.0
unattended-upgrades==0.1
urllib3==1.26.12
